requirements.txt
requests
undetected-chromedriver
selenium
Pillow
rich